pyproject.toml
[project]
name = "tj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "beautifulsoup4>=4.14.2",
    "django>=5.2.8",
    "langchain>=1.1.0",
    "langchain-ollama>=1.0.0",
    "markdown>=3.10",
    "pdfplumber>=0.11.8",
    "pgvector>=0.4.1",
    "psycopg>=3.2.13",
    "psycopg-binary>=3.2.13",
    "python-dotenv>=1.2.1",
    "sendgrid>=6.12.5",
    "sentence-transformers>=5.1.2",
]
